Split codex op handlers
